feat: support multiple component paths
Fixes #1.

Adds support for the `COMPONENT_PATH` environment variable, which
is a path to one or more component directories, defaulting it to
the built-in `components` directory if none is given. Then updates
create and deploy flows to search this path for the component.

Removes support for COMPONENT_BASE_DIR environment variable.

Together these support the inclusion of this tool as a dependency.